% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_student_repos.R
\name{create_student_repos}
\alias{create_student_repos}
\title{Create Student Repositories}
\usage{
create_student_repos(
  assignment_name,
  org_name = "ma22019-2026",
  roster_path = "../roster_public.csv"
)
}
\arguments{
\item{assignment_name}{String, e.g. "hw_01" or "cw_01"}

\item{org_name}{String, default "ma22019-2026"}

\item{roster_path}{String, path to CSV roster}
}
\description{
Reads the roster and creates private repositories for each student
in the specified organization. Adds the student as a collaborator.
}
